api/pyproject.toml
[tool.poetry]
name = "tgfeed"
version = "0.1.0"
description = "TBD"
authors = ["slay_lines"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
telethon = "^1.10"
toml = "^0.10.0"
pymongo = "^3.9"
PySocks = "^1.7"
fastapi = "^0.42.0"
boto3 = "^1.11"
cryptg = "^0.2.0"
python-magic = "^0.4.15"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
